Move the nomis models out of services
Currently the models for NOMIS objects are stored in
app/services/models, but we already have a models folder.

This PR will move all of the models from services/NOMIS::MODELS to
models/NOMIS::
