+++ app/jobs/movement_job.rb
@@ -2,6 +2,6 @@ class MovementJob < ApplicationJob
   queue_as :default
 
   def perform(mvment)
-    MovementService.process_movement(Nomis::Models::Movement.new(JSON.parse(mvment)))
+    MovementService.process_movement(Nomis::Movement.new(JSON.parse(mvment)))
   end
 end
