+++ app/models/nomis/null_keyworker.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Nomis
+  class NullKeyworker < KeyworkerDetails
+    def full_name
+      'Data not available'
+    end
+  end
+end
