--- app/services/nomis/models/keyworker_details.rb
@@ -1,23 +1,0 @@
-# frozen_string_literal: true
-
-module Nomis
-  module Models
-    class KeyworkerDetails
-      include Deserialisable
-
-      attr_accessor :staff_id, :first_name, :last_name
-
-      def self.from_json(json)
-        KeyworkerDetails.new.tap { |obj|
-          obj.staff_id = json['staffId']&.to_i
-          obj.first_name = json['firstName']
-          obj.last_name = json['lastName']
-        }
-      end
-
-      def full_name
-        "#{last_name}, #{first_name}".titleize
-      end
-    end
-  end
-end
