--- app/services/nomis/models/null_keyworker.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Nomis
-  module Models
-    class NullKeyworker < KeyworkerDetails
-      def full_name
-        'Data not available'
-      end
-    end
-  end
-end
