+++ spec/helpers/application_helper_spec.rb
@@ -24,15 +24,15 @@ RSpec.describe ApplicationHelper do
 
   describe 'generates labels for case owner ' do
     it 'can show Custody for Prison' do
-      off = Nomis::Models::Offender.new
-      off.sentence = Nomis::Models::SentenceDetail.new
+      off = Nomis::Offender.new
+      off.sentence = Nomis::SentenceDetail.new
 
       expect(case_owner_label(off)).to eq('Custody')
     end
 
     it 'can show Community for Probation' do
-      off = Nomis::Models::Offender.new
-      off.sentence = Nomis::Models::SentenceDetail.new
+      off = Nomis::Offender.new
+      off.sentence = Nomis::SentenceDetail.new
       off.sentence.release_date = Time.zone.today
 
       expect(case_owner_label(off)).to eq('Community')
