+++ spec/helpers/search_helper_spec.rb
@@ -3,7 +3,7 @@ require 'rails_helper'
 RSpec.describe SearchHelper do
   describe 'the CTA' do
     it "will change to edit if there is no tier" do
-      offender = Nomis::Models::Offender.new(
+      offender = Nomis::Offender.new(
         offender_no: 'A'
       )
       text, _link = cta_for_offender('LEI', offender)
@@ -11,7 +11,7 @@ RSpec.describe SearchHelper do
     end
 
     it "will change to allocate if there is no allocation" do
-      offender = Nomis::Models::Offender.new(
+      offender = Nomis::Offender.new(
         offender_no: 'A',
         tier: 'A'
       )
@@ -20,7 +20,7 @@ RSpec.describe SearchHelper do
     end
 
     it "will change to view if there is an allocation" do
-      offender = Nomis::Models::Offender.new(
+      offender = Nomis::Offender.new(
         offender_no: 'A',
         tier: 'A',
         allocated_pom_name: 'Bob'
