+++ spec/services/bucket_spec.rb
@@ -31,9 +31,9 @@ describe Bucket do
 
   it "can sort its items by default ASC" do
     b = described_class.new
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "Z" }
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "A" }
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "M" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "Z" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "A" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "M" }
     expect(b.items.count).to be(3)
 
     b.sort(:last_name)
@@ -44,9 +44,9 @@ describe Bucket do
 
   it "can sort its items DESC" do
     b = described_class.new
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "A" }
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "M" }
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "Z" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "A" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "M" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "Z" }
     expect(b.items.count).to be(3)
 
     b.sort(:last_name, :desc)
@@ -57,9 +57,9 @@ describe Bucket do
 
   it "can't sort by made up field" do
     b = described_class.new
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "A" }
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "M" }
-    b << Nomis::Models::OffenderSummary.new.tap { |o| o.last_name = "Z" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "A" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "M" }
+    b << Nomis::OffenderSummary.new.tap { |o| o.last_name = "Z" }
     expect(b.items.count).to be(3)
 
     b.sort(:test, :desc)
@@ -70,13 +70,13 @@ describe Bucket do
 
   it 'can sort dates containing nulls' do
     b = described_class.new
-    b << Nomis::Models::OffenderSummary.new.tap { |o|
+    b << Nomis::OffenderSummary.new.tap { |o|
       o.last_name = "A"
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
     }
-    b << Nomis::Models::OffenderSummary.new.tap { |o|
+    b << Nomis::OffenderSummary.new.tap { |o|
       o.last_name = "A"
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.release_date = Date.new(2000, 1, 1)
     }
     b.sort(:earliest_release_date)
