+++ spec/services/movement_service_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
-require_relative '../../app/services/nomis/models/movement'
+require_relative '../../app/models/nomis/movement'
 
 describe MovementService do
   let!(:new_offender_court) { create(:movement, offender_no: 'G4273GI', from_agency: 'COURT')   }
@@ -11,14 +11,14 @@ describe MovementService do
     movements = described_class.movements_on(Date.iso8601('2019-02-20'))
     expect(movements).to be_kind_of(Array)
     expect(movements.length).to eq(2)
-    expect(movements.first).to be_kind_of(Nomis::Models::Movement)
+    expect(movements.first).to be_kind_of(Nomis::Movement)
   end
 
   it "can filter transfer type results",
      vcr: { cassette_name: :movement_service_filter_type_spec }  do
     movements = described_class.movements_on(
       Date.iso8601('2019-02-20'),
-      type_filters: [Nomis::Models::MovementType::TRANSFER]
+      type_filters: [Nomis::MovementType::TRANSFER]
     )
     expect(movements).to be_kind_of(Array)
     expect(movements.length).to eq(0)
@@ -28,48 +28,48 @@ describe MovementService do
      vcr: { cassette_name: :movement_service_filter_adm_spec }  do
     movements = described_class.movements_on(
       Date.iso8601('2019-03-12'),
-      type_filters: [Nomis::Models::MovementType::ADMISSION]
+      type_filters: [Nomis::MovementType::ADMISSION]
     )
     expect(movements).to be_kind_of(Array)
     expect(movements.length).to eq(1)
-    expect(movements.first).to be_kind_of(Nomis::Models::Movement)
-    expect(movements.first.movement_type).to eq(Nomis::Models::MovementType::ADMISSION)
+    expect(movements.first).to be_kind_of(Nomis::Movement)
+    expect(movements.first.movement_type).to eq(Nomis::MovementType::ADMISSION)
   end
 
   it "can filter release type results",
      vcr: { cassette_name: :movement_service_filter_release_spec }  do
     movements = described_class.movements_on(
       Date.iso8601('2019-02-20'),
-      type_filters: [Nomis::Models::MovementType::RELEASE]
+      type_filters: [Nomis::MovementType::RELEASE]
     )
     expect(movements).to be_kind_of(Array)
     expect(movements.length).to eq(2)
-    expect(movements.first).to be_kind_of(Nomis::Models::Movement)
-    expect(movements.first.movement_type).to eq(Nomis::Models::MovementType::RELEASE)
+    expect(movements.first).to be_kind_of(Nomis::Movement)
+    expect(movements.first.movement_type).to eq(Nomis::MovementType::RELEASE)
   end
 
   it "can filter results by direction IN",
      vcr: { cassette_name: :movement_service_filter_direction_in_spec }  do
     movements = described_class.movements_on(
       Date.iso8601('2019-03-12'),
-      direction_filters: [Nomis::Models::MovementDirection::IN]
+      direction_filters: [Nomis::MovementDirection::IN]
     )
     expect(movements).to be_kind_of(Array)
     expect(movements.length).to eq(1)
-    expect(movements.first).to be_kind_of(Nomis::Models::Movement)
-    expect(movements.first.direction_code).to eq(Nomis::Models::MovementDirection::IN)
+    expect(movements.first).to be_kind_of(Nomis::Movement)
+    expect(movements.first.direction_code).to eq(Nomis::MovementDirection::IN)
   end
 
   it "can filter results by direction OUT",
      vcr: { cassette_name: :movement_service_filter_direction_out_spec }  do
     movements = described_class.movements_on(
       Date.iso8601('2019-02-20'),
-      direction_filters: [Nomis::Models::MovementDirection::OUT]
+      direction_filters: [Nomis::MovementDirection::OUT]
     )
     expect(movements).to be_kind_of(Array)
     expect(movements.length).to eq(2)
-    expect(movements.first).to be_kind_of(Nomis::Models::Movement)
-    expect(movements.first.direction_code).to eq(Nomis::Models::MovementDirection::OUT)
+    expect(movements.first).to be_kind_of(Nomis::Movement)
+    expect(movements.first.direction_code).to eq(Nomis::MovementDirection::OUT)
   end
 
   it "can ignore movements OUT",
@@ -132,7 +132,7 @@ describe MovementService do
 
     it "will not process offenders on remand",
        vcr: { cassette_name: :movement_service_transfer_in__not_convicted_spec }  do
-      allow(OffenderService).to receive(:get_offender).and_return(Nomis::Models::Offender.new.tap{ |o|
+      allow(OffenderService).to receive(:get_offender).and_return(Nomis::Offender.new.tap{ |o|
         o.convicted_status = "Remand"
       })
       processed = described_class.process_movement(transfer_in)
