+++ spec/services/offender_service_spec.rb
@@ -8,14 +8,14 @@ describe OffenderService do
     offenders = described_class.get_offenders_for_prison('LEI').first(9)
     expect(offenders).to be_kind_of(Array)
     expect(offenders.length).to eq(9)
-    expect(offenders.first).to be_kind_of(Nomis::Models::OffenderSummary)
+    expect(offenders.first).to be_kind_of(Nomis::OffenderSummary)
   end
 
   it "get last page of offenders for a specific prison", vcr: { cassette_name: :offender_service_offenders_by_prison_last_page_spec } do
     offenders = described_class.get_offenders_for_prison('LEI').to_a
     expect(offenders).to be_kind_of(Array)
     expect(offenders.length).to eq(823)
-    expect(offenders.first).to be_kind_of(Nomis::Models::OffenderSummary)
+    expect(offenders.first).to be_kind_of(Nomis::OffenderSummary)
   end
 
   it "gets a single offender", vcr: { cassette_name: :offender_service_single_offender_spec } do
@@ -24,7 +24,7 @@ describe OffenderService do
     create(:case_information, nomis_offender_id: nomis_offender_id, tier: 'C', case_allocation: 'CRC', omicable: 'Yes')
     offender = described_class.get_offender(nomis_offender_id)
 
-    expect(offender).to be_kind_of(Nomis::Models::Offender)
+    expect(offender).to be_kind_of(Nomis::Offender)
     expect(offender.sentence.release_date).to eq Date.new(2020, 2, 7)
     expect(offender.tier).to eq 'C'
     expect(offender.main_offence).to eq 'Section 18 - wounding with intent to resist / prevent arrest'
@@ -52,7 +52,7 @@ describe OffenderService do
 
       updated_offenders = described_class.set_allocated_pom_name(offenders, 'LEI')
       expect(updated_offenders).to be_kind_of(Array)
-      expect(updated_offenders.first).to be_kind_of(Nomis::Models::OffenderSummary)
+      expect(updated_offenders.first).to be_kind_of(Nomis::OffenderSummary)
       expect(updated_offenders.count).to eq(offenders.count)
       expect(updated_offenders.first.allocated_pom_name).to eq('Jones, Ross')
       expect(updated_offenders.first.allocation_date).to be_kind_of(Date)
