+++ spec/services/responsibility_service_spec.rb
@@ -2,91 +2,91 @@ require 'rails_helper'
 
 describe ResponsibilityService do
   let(:offender_no_release_date) {
-    Nomis::Models::Offender.new.tap { |o|
-      o.sentence = Nomis::Models::SentenceDetail.new
+    Nomis::Offender.new.tap { |o|
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.release_date = nil
     }
   }
 
   let(:offender_not_welsh) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = false
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.release_date = DateTime.now.utc.to_date + 6.months
     }
   }
 
   let(:offender_welsh_crc_lt_12_wk) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = true
       o.case_allocation = 'CRC'
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.release_date = DateTime.now.utc.to_date + 2.weeks
     }
   }
 
   let(:offender_welsh_crc_gt_12_wk) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = true
       o.case_allocation = 'CRC'
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.release_date = DateTime.now.utc.to_date + 13.weeks
     }
   }
 
   let(:offender_welsh_nps_gt_10_mths) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = true
       o.case_allocation = 'NPS'
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.release_date = DateTime.now.utc.to_date + 11.months
     }
   }
 
   let(:offender_welsh_nps_lt_10_mths) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = true
       o.case_allocation = 'NPS'
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.release_date = DateTime.now.utc.to_date + 9.months
     }
   }
 
   let(:offender_welsh_nps_old_case_gt_15_mths) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = true
       o.case_allocation = 'NPS'
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.sentence_start_date = DateTime.new(2019, 1, 19).utc
       o.sentence.release_date = DateTime.now.utc.to_date + 16.months
     }
   }
 
   let(:offender_welsh_nps_old_case_lt_15_mths) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = true
       o.case_allocation = 'NPS'
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.sentence_start_date = DateTime.new(2019, 2, 20).utc
       o.sentence.release_date = DateTime.now.utc.to_date + 9.months
     }
   }
 
   let(:offender_welsh_nps_new_case_gt_10_mths) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = true
       o.case_allocation = 'NPS'
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.sentence_start_date = DateTime.new(2019, 1, 19).utc
       o.sentence.release_date = DateTime.now.utc.to_date + 16.months
     }
   }
 
   let(:offender_welsh_nps_new_case_lt_10_mths) {
-    Nomis::Models::Offender.new.tap { |o|
+    Nomis::Offender.new.tap { |o|
       o.omicable = true
       o.case_allocation = 'NPS'
-      o.sentence = Nomis::Models::SentenceDetail.new
+      o.sentence = Nomis::SentenceDetail.new
       o.sentence.sentence_start_date = DateTime.new(2019, 2, 20).utc
       o.sentence.release_date = DateTime.now.utc.to_date + 9.months
     }
